Week7/Day5/MiniProjet/genai_pipeline_excellence/requirements.txt
transformers>=4.45.0
sentence-transformers>=3.0.1
torch>=2.2.0
scikit-learn>=1.5.0
numpy>=1.26.0
scipy>=1.11.0
regex>=2023.10.3
python-Levenshtein>=0.23.0
textstat>=0.7.4
rouge-score>=0.1.2
sacrebleu>=2.4.0
streamlit>=1.38.0
schedule>=1.2.1
PyYAML>=6.0.1
sqlite-utils>=3.37
